Ft support discord

+++ index.js
@@ -15,18 +15,30 @@ console.log();
 
 let slack = `${user}\\AppData\\Roaming\\Slack\\Cache`;
 let chrome = `${user}\\AppData\\Local\\Google\\Chrome\\User Data\\Default\\Cache`;
+let discord = `${user}/AppData/Roaming/discord/Cache`;
+let discord_canary = `${user}/AppData/Roaming/discordcanary/Cache`;
+let discord_ptb = `${user}/AppData/Roaming/discordptb/Cache`;
 
 if (os.platform() === "linux") {
   chrome = `${user}/.cache/google-chrome/Default/Cache/`;
   slack = `${user}/.config/Slack/Cache`;
+  discord = `${user}/.config/discord/Cache`;
+  discord_canary = `${user}/.config/discordcanary/Cache`;
+  discord_ptb = `${user}/.config/discordptb/Cache`;
 }
 if (os.platform() === "darwin") {
   chrome = `${user}/Library/Application\ Support/Google/Chrome/Default/Application\ Cache/Cache`;
   slack = `${user}/Library/Application Support/Slack/Cache`;
+  discord = `${user}/Library/Application Support/discord/Cache`;
+  discord_canary = `${user}/Library/Application Support/discordcanary/Cache`;
+  discord_ptb = `${user}/Library/Application Support/discordptb/Cache`;
 }
 const cachesDirs = [
   { name: "slack", path: slack },
   { name: "chrome", path: chrome },
+  { name: "discord", path: discord },
+  { name: "discord_canary", path: discord_canary },
+  { name: "discord_ptb", path: discord_ptb },
 ];
 // get cached media size
 const cacheChecker = (dir) => {
@@ -41,7 +53,9 @@ const cacheChecker = (dir) => {
 const getTotalSize = () => {
   let total = 0;
   cachesDirs.forEach((cacheDir) => {
-    total += cacheChecker(cacheDir.path);
+    if (fs.existsSync(cacheDir.path)) {
+      total += cacheChecker(cacheDir.path);
+    }
   });
   if (total.toString().length > 6)
     return Math.floor(total / (1024 * 1024)).toString() + " Mbs";
@@ -52,10 +66,12 @@ const getTotalSize = () => {
 const deleteCache = () => {
   try {
     cachesDirs.forEach((cacheDir) => {
-      const files = fs.readdirSync(cacheDir.path);
-      files.forEach((file) => {
-        fs.unlinkSync(cacheDir.path + "\\" + file);
-      });
+      if (fs.existsSync(cacheDir.path)) {
+        const files = fs.readdirSync(cacheDir.path);
+        files.forEach((file) => {
+          fs.unlinkSync(cacheDir.path + "\\" + file);
+        });
+      }
     });
     spinner.succeed("Done");
   } catch (error) {
